overlay: add new module for installing Azure udev rules
We want the Azure udev rules present in the initrd, so that Ignition
is able to detect the disks and act on them.

If the udev rules end up being installed into the initramfs as part of
the WALinuxAgent-udev package, the dracut module should be removed.

See https://bugzilla.redhat.com/show_bug.cgi?id=1909287